Fix Dreamcast portability and memory safety
Improves Dreamcast compatibility across file I/O, directory handling, PhysFS usage, input mapping, rumble, and platform-specific constants/includes. Hardens core memory paths by adding aligned arena allocation, overflow/range checks, safer realloc/header validation, and explicit extra-pool allocation entry points. Also refactors MD5 hashing into an update/finish context API, fixes WAV ADPCM cleanup and array deletes, tightens numeric formatting/casts, and removes an unused string replace helper.

+++ epi/fxp_bbox.cc
@@ -27,16 +27,16 @@ namespace epi
 std::string xbox2_c::ToStr(int precision) const
 {
 	return STR_Format("[%1.*f:%1.*f %1.*f:%1.*f]",
-			precision, lo.x.ToFloat(), precision, hi.x.ToFloat(),
-			precision, lo.y.ToFloat(), precision, hi.y.ToFloat());
+			precision, static_cast<double>(lo.x.ToFloat()), precision, static_cast<double>(hi.x.ToFloat()),
+			precision, static_cast<double>(lo.y.ToFloat()), precision, static_cast<double>(hi.y.ToFloat()));
 }
 
 std::string xbox3_c::ToStr(int precision) const
 {
 	return STR_Format("[%1.*f:%1.*f %1.*f:%1.*f %1.*f:%1.*f]",
-			precision, lo.x.ToFloat(), precision, hi.x.ToFloat(),
-			precision, lo.y.ToFloat(), precision, hi.y.ToFloat(),
-			precision, lo.z.ToFloat(), precision, hi.z.ToFloat());
+			precision, static_cast<double>(lo.x.ToFloat()), precision, static_cast<double>(hi.x.ToFloat()),
+			precision, static_cast<double>(lo.y.ToFloat()), precision, static_cast<double>(hi.y.ToFloat()),
+			precision, static_cast<double>(lo.z.ToFloat()), precision, static_cast<double>(hi.z.ToFloat()));
 }
 
 } // namespace epi
